% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_bracket.R
\name{bracket_draw_left}
\alias{bracket_draw_left}
\title{Draw bracket match}
\usage{
bracket_draw_left(x, y, l, w)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{l}{length of line}

\item{w}{width between two teams lines}
}
\value{
returns a dataframe with coordinates to draw a match
}
\description{
This function is a helper function which draws one match on a bracket
}
