% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-api.R
\name{get_tournament_scores_api}
\alias{get_tournament_scores_api}
\title{Get live scores from NCAA tournament}
\usage{
get_tournament_scores_api(league = "mens", year = NULL)
}
\arguments{
\item{league}{"mens" or "womens"}

\item{year}{NCAA tournament year}
}
\value{
data.frame with 63 games and scores.
}
\description{
Get live scores from NCAA tournament for auction and report app scoring.
}
