% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa-data.r
\name{import_teams}
\alias{import_teams}
\title{Import csv of NCAA teams}
\usage{
import_teams(
  file_path = system.file("extdata", "ncaa-teams.csv", package = "NCAAcalcutta", mustWork
    = TRUE)
)
}
\description{
Import csv of NCAA teams
}
\examples{
\dontrun{
# Input the 2018 teams
teams <- import_teams(system.file("extdata", "ncaa-teams.csv", package = "NCAAcalcutta"))
}

}
