\name{ncaa-calcutta-package}
\alias{ncaa-calcutta-package}
\alias{ncaa-calcutta}
\docType{package}
\title{
\packageTitle{ncaa-calcutta}
}
\description{
\packageDescription{ncaa-calcutta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ncaa-calcutta}
\packageIndices{ncaa-calcutta}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ncaa-calcutta}

Maintainer: \packageMaintainer{ncaa-calcutta}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
