% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa-data.r
\name{randomize_teams}
\alias{randomize_teams}
\title{Randomize order of teams for Calcutta auction}
\usage{
randomize_teams(df, random_seed = 156)
}
\description{
Randomize order of teams for Calcutta auction
}
\examples{
\dontrun{
# Input the 2018 teams
teams <- import_teams(system.file("extdata", "ncaa-teams.csv", package = "NCAAcalcutta"))
teams <- randomize_teams(teams, random_seed=123)
}

}
