% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{scrape_teams}
\alias{scrape_teams}
\title{Scrape the team names and ids from the ESPN NCAA MBB index}
\usage{
scrape_teams(league)
}
\arguments{
\item{league}{either 'mens' or 'womens'}
}
\value{
data.frame of team names and ids
}
\description{
Scrape the team names and ids from the ESPN NCAA MBB index
}
\author{
eshayer
}
