% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auction-app.R
\name{start_auction}
\alias{start_auction}
\title{Start the NCAA Calcutta Shiny App}
\usage{
start_auction(teams, randomize = TRUE)
}
\description{
Start the NCAA Calcutta Shiny App
}
\examples{
# Input the 2018 teams
teams <- import_teams(system.file("extdata", "ncaa-teams.csv", package = "NCAAcalcutta"))
start_auction(teams, randomize=TRUE)

}
