% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa-data.r
\name{team_skeleton}
\alias{team_skeleton}
\title{Create a skeleton csv to manually input teams}
\usage{
team_skeleton()
}
\description{
Create a skeleton csv to manually input teams
}
\examples{
\dontrun{
# Prep file to input teams manually
preteams <- team_skeleton()
}

}
