% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_bracket.R
\name{tournament_bracket}
\alias{tournament_bracket}
\title{Draw tournament bracket}
\usage{
tournament_bracket(n_teams, w = 0.5, l = 5)
}
\arguments{
\item{n_teams}{Number of teams playing in the tournament (should be a multiple of 4)}

\item{w}{width between two teams lines}

\item{l}{length of line}
}
\value{
plotly object with drawn bracket
}
\description{
Draw tournament bracket
}
