% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auction-app.R
\name{start_auction}
\alias{start_auction}
\title{Start the NCAA Calcutta Shiny App}
\usage{
start_auction(teams, players, points, randomize = TRUE)
}
\arguments{
\item{teams}{This should be a dataframe with rank, region, team, and opponent}

\item{players}{This should be a character vector of all those participating}

\item{points}{This is the total number of points each player starts with}

\item{randomize}{(Default: TRUE) This determines whether you'd like to randomize the order of which teams are auctioned
By default this is a block randomization so that you will get a 1-16 seed before you see the next group of 1-16 seeds}
}
\description{
This Shiny App will help facilitate the auction
}
\examples{
\dontrun{
# Input the 2018 teams
teams <- import_teams(system.file("extdata", "ncaa-teams.csv", package = "NCAAcalcutta"))
players <- c('Mark','Markus','Marko','Marky')
points <- 1200
start_auction(teams, players, points, randomize=TRUE)
}

}
